refactoring: move papers.Paper to the new package
