+++ main.go
@@ -29,9 +29,7 @@ import (
 	"fmt"
 	"io"
 	"log"
-	"net/url"
 	"os"
-	"regexp"
 	"sort"
 	"strings"
 	"text/template"
@@ -40,8 +38,8 @@ import (
 	"unicode/utf8"
 
 	"github.com/bzz/scholar-alert-digest/gmailutils"
+	"github.com/bzz/scholar-alert-digest/papers"
 
-	"github.com/antchfx/htmlquery"
 	"gitlab.com/golang-commonmark/markdown"
 	"google.golang.org/api/gmail/v1"
 )
@@ -105,8 +103,7 @@ The -subj flag will only include email subjects in the report. Usefull for pipin
 )
 
 var (
-	user             = "me"
-	scholarURLPrefix = regexp.MustCompile(`http(s)?://scholar\.google\.\p{L}+/scholar_url\?url=`)
+	user = "me"
 
 	gmailLabel = flag.String("l", labelName, "name of the Gmail label")
 	listLabels = flag.Bool("labels", false, "list all Gmail labels")
@@ -143,7 +140,11 @@ func main() {
 	}
 
 	if *onlySubj {
-		msgs := gmailutils.Fetch(srv, user, fmt.Sprintf("label:%s", *gmailLabel))
+		msgs, err := gmailutils.Fetch(context.TODO(), srv, user, fmt.Sprintf("label:%s", *gmailLabel))
+		if err != nil {
+			log.Fatal("Failed to fetch messages from Gmail: %v", err)
+		}
+
 		printSubjects(msgs)
 		os.Exit(0)
 	}
@@ -151,17 +152,17 @@ func main() {
 	// TODO(bzz): FetchAsync returning chan *gmail.Message
 	urMsgs, err := gmailutils.Fetch(context.TODO(), srv, user, fmt.Sprintf("label:%s is:unread", *gmailLabel))
 	if err != nil {
-		log.Fatal("Failed to fetch messages from Gmail")
+		log.Fatal("Failed to fetch messages from Gmail: %v", err)
 	}
-	errCnt, urTitlesCnt, urTitles := extractPapersFromMsgs(urMsgs)
+	errCnt, urTitlesCnt, urTitles := papers.ExtractPapersFromMsgs(urMsgs)
 
-	var rTitles map[paper]int
+	var rTitles map[papers.Paper]int
 	if *read {
 		rMsgs, err := gmailutils.Fetch(context.TODO(), srv, user, fmt.Sprintf("label:%s is:read", *gmailLabel))
 		if err != nil {
 			log.Fatal("Failed to fetch messages from Gmail")
 		}
-		_, _, rTitles = extractPapersFromMsgs(rMsgs)
+		_, _, rTitles = papers.ExtractPapersFromMsgs(rMsgs)
 	}
 
 	if *ouputHTML {
@@ -217,116 +218,7 @@ func splitOnDash(str string) ([]string, string) {
 	return strings.Split(str, sep), sep
 }
 
-func extractPapersFromMsgs(messages []*gmail.Message) (int, int, map[paper]int) {
-	errCount := 0
-	titlesCount := 0
-	uniqTitles := map[paper]int{}
-
-	for _, m := range messages {
-		papers, err := extractPapersFromMsg(m)
-		if err != nil {
-			errCount++
-			continue
-		}
-
-		titlesCount += len(papers)
-		for _, paper := range papers { // map title to uniqTitles
-			uniqTitles[paper]++
-		}
-	}
-
-	return errCount, titlesCount, uniqTitles
-}
-
-func extractPapersFromMsg(m *gmail.Message) ([]paper, error) {
-	subj := gmailutils.Subject(m.Payload)
-
-	body, err := gmailutils.MessageTextBody(m)
-	if err != nil {
-		e := fmt.Errorf("failed to get message text for ID %s - %s", m.Id, err)
-		return nil, e
-	}
-
-	doc, err := htmlquery.Parse(bytes.NewReader(body))
-	if err != nil {
-		e := fmt.Errorf("failed to parse HTML body of %q", subj)
-		return nil, e
-	}
-
-	// paper titles, from a single email
-	xpTitle := "//h3/a"
-	titles, err := htmlquery.QueryAll(doc, xpTitle)
-	if err != nil {
-		return nil, fmt.Errorf("title: not valid XPath expression %q", xpTitle)
-	}
-
-	// paper urls, from a single email
-	xpURL := "//h3/a/@href"
-	urls, err := htmlquery.QueryAll(doc, xpURL)
-	if err != nil {
-		return nil, fmt.Errorf("url: not valid XPath expression %q", xpURL)
-	}
-
-	if len(titles) != len(urls) {
-		e := fmt.Errorf("titles %d != %d urls in %q", len(titles), len(urls), subj)
-		return nil, e
-	}
-
-	// paper abstract
-	xpAbs := "//h3/following-sibling::div[2]"
-	abss, err := htmlquery.QueryAll(doc, xpAbs)
-	if err != nil {
-		return nil, fmt.Errorf("abstract: not valid XPath expression %q", xpAbs)
-	}
-
-	var papers []paper
-	for i, aTitle := range titles {
-		title := strings.TrimSpace(htmlquery.InnerText(aTitle))
-		abs := strings.TrimSpace(htmlquery.InnerText(abss[i]))
-
-		url, err := extractPaperURL(htmlquery.InnerText(urls[i]))
-		if err != nil {
-			log.Printf("Skipping paper %q in %q: %s", title, subj, err)
-			continue
-		}
-
-		papers = append(papers, paper{
-			title, url, abstract{
-				abs, separateFirstLine(abs)[0], separateFirstLine(abs)[1],
-			},
-		})
-	}
-	return papers, nil
-}
-
-// extractPaperURL returns an actual paper URL from the given scholar link.
-// Does not validate URL format but extracts it ad-hoc by trimming sufix/prefix.
-func extractPaperURL(scholarURL string) (string, error) {
-	// drop scholarURLPrefix
-	prefixLoc := scholarURLPrefix.FindStringIndex(scholarURL)
-	if prefixLoc == nil {
-		return "", fmt.Errorf("url %q does not have prefix %q", scholarURL, scholarURLPrefix.String())
-	}
-	longURL := scholarURL[prefixLoc[1]:]
-
-	// drop sufix (after &), if any
-	if sufix := strings.Index(longURL, "&"); sufix >= 0 {
-		longURL = longURL[:sufix]
-	}
-
-	return url.QueryUnescape(longURL)
-}
-
-func separateFirstLine(text string) []string {
-	text = strings.ReplaceAll(text, "\n", "")
-	n := 80 // TODO(bzz): utf8 whitespace-aware splitting alg capped by max N runes
-	if len(text) < n {
-		return []string{text, ""}
-	}
-	return []string{text[:n], text[n:]}
-}
-
-func generateAndPrintHTML(msgsCnt, titlesCnt int, unread, read map[paper]int) {
+func generateAndPrintHTML(msgsCnt, titlesCnt int, unread, read map[papers.Paper]int) {
 	var mdBuf bytes.Buffer
 	generateMarkdown(&mdBuf, msgsCnt, titlesCnt, unread, read)
 
@@ -334,11 +226,11 @@ func generateAndPrintHTML(msgsCnt, titlesCnt int, unread, read map[paper]int) {
 	fmt.Printf(htmlTemplText, md.RenderToString([]byte(mdBuf.String())))
 }
 
-func generateAndPrintMarkdown(msgsCnt, titlesCnt int, unread, read map[paper]int) {
+func generateAndPrintMarkdown(msgsCnt, titlesCnt int, unread, read map[papers.Paper]int) {
 	generateMarkdown(os.Stdout, msgsCnt, titlesCnt, unread, read)
 }
 
-func generateMarkdown(out io.Writer, msgsCnt, titlesCnt int, unread, read map[paper]int) {
+func generateMarkdown(out io.Writer, msgsCnt, titlesCnt int, unread, read map[papers.Paper]int) {
 	newMdReport(out, msgsCnt, titlesCnt, unread)
 	if read != nil {
 		oldMdReport(out, read)
@@ -346,23 +238,23 @@ func generateMarkdown(out io.Writer, msgsCnt, titlesCnt int, unread, read map[pa
 }
 
 // renderes newMdTemplText for new, unread papers.
-func newMdReport(out io.Writer, msgsCnt, titlesCnt int, papers map[paper]int) {
+func newMdReport(out io.Writer, msgsCnt, titlesCnt int, agrPapers map[papers.Paper]int) {
 	tmplText := newMdTemplText
 	tmpl := template.Must(template.New("unread-papers").Funcs(template.FuncMap{
-		"sortedKeys": sortedKeys,
+		"sortedKeys": papers.SortedKeys,
 	}).Parse(tmplText))
 	err := tmpl.Execute(out, struct {
 		Date         string
 		UnreadEmails int
 		TotalPapers  int
 		UniqPapers   int
-		Papers       map[paper]int
+		Papers       map[papers.Paper]int
 	}{
 		time.Now().Format(time.RFC3339),
 		msgsCnt,
 		titlesCnt,
-		len(papers),
-		papers,
+		len(agrPapers),
+		agrPapers,
 	})
 	if err != nil {
 		log.Fatalf("template %q execution failed: %s", tmplText, err)
@@ -370,12 +262,12 @@ func newMdReport(out io.Writer, msgsCnt, titlesCnt int, papers map[paper]int) {
 }
 
 // renderes oldMdTemplText for old, read papers
-func oldMdReport(out io.Writer, papers map[paper]int) {
+func oldMdReport(out io.Writer, agrPapers map[papers.Paper]int) {
 	tmplText := oldMdTemplText
 	tmpl := template.Must(template.New("read-papers").Funcs(template.FuncMap{
-		"sortedKeys": sortedKeys,
+		"sortedKeys": papers.SortedKeys,
 	}).Parse(tmplText))
-	err := tmpl.Execute(out, papers)
+	err := tmpl.Execute(out, agrPapers)
 	if err != nil {
 		log.Fatalf("template %q execution failed: %s", tmplText, err)
 	}
@@ -398,42 +290,4 @@ func markGmailMsgsUnread(srv *gmail.Service, user string, messages []*gmail.Mess
 	}
 	// TODO(bzz): move to
 	//  gmailutils.ModifyMessagesDelLabel(srv, user, messages, "UNREAD")
-}
-
-// Helpers for a Map, sorted by keys.
-// TODO(bzz): move to map.go after `go run main.go` is replaced by ./cmd/report
-type sortedMap struct {
-	m map[paper]int
-	s []paper
-}
-
-func (sm *sortedMap) Len() int           { return len(sm.m) }
-func (sm *sortedMap) Less(i, j int) bool { return sm.m[sm.s[i]] > sm.m[sm.s[j]] }
-func (sm *sortedMap) Swap(i, j int)      { sm.s[i], sm.s[j] = sm.s[j], sm.s[i] }
-
-// TODO(bzz): use a stable sort
-func sortedKeys(m map[paper]int) []paper {
-	sm := new(sortedMap)
-	sm.m = m
-	sm.s = make([]paper, len(m))
-	i := 0
-	for key := range m {
-		sm.s[i] = key
-		i++
-	}
-	sort.Sort(sm)
-	return sm.s
-}
-
-// paper is a map key, thus aggregation take into account all it's fields.
-//
-// TODO(bzz): think about aggregation only by the title, as suggested in
-// https://github.com/bzz/scholar-alert-digest/issues/12#issuecomment-562820924
-type paper struct {
-	Title, URL string
-	Abstract   abstract
-}
-
-type abstract struct {
-	Full, FirstLine, RestLines string
 }
